Sat, Oct 21, 2023, 4:41 PM +05:30

+++ half/half.c
@@ -18,8 +18,8 @@ int main(void)
 // TODO: Complete the function
 float half(float bill, float tax, int tip)
 {
-    tax = (float)(tax*bill)/100;
-    tip = (float)(tip*(tax+bill))/100;
+    tax = (tax*bill)/100;
+    tip =tip*(tax+bill)/100;
     float half_amount = (bill+tax+tip)/2;
     return half_amount;
